refactor: refactor keyButton
