--- src/components/KeyButton.tsx
@@ -1,13 +1,0 @@
-interface KeyProps {
-  className: string;
-  onClick: () => void;
-  label: string;
-}
-
-export default function KeyButton({ className, onClick, label }: KeyProps) {
-  return (
-    <button className={className} onClick={onClick}>
-      {label}
-    </button>
-  );
-}
